fix(canvas): preset panel の outside-click 判定を親 ref へ移管 + コメント微修正 (#522)
- TeamPresetsPanel.tsx の `wrapRef` ベースの pointerdown ハンドラを撤去
  - 子側で持つと、トグルボタンの pointerdown が "外側→close" と判定されたあと
    続く onClick で再 open される競合 (open→close→open) を起こしていた
- StageHud.tsx の `presetsWrapRef` (ボタン + popover を内包) を ref 対象に
  外クリック / Escape ハンドラを追加。arrange popover と同じパターンに揃える
- canvas.css のコメント「長 preset 名」を「長い preset 名」に修正
